organize 'parser-factory' a bit better
