+++ packages/parser-factory/src/builders.ts
@@ -99,8 +99,6 @@ export const pow = (
     loc,
 });
 
-// NOTE: we don't use a default param here since we want individual
-// nodes to be created for the index of each root.
 export const root = (
     radicand: types.Node,
     index: types.Node,
